+++ src/main/joyride/db.cljs
@@ -3,7 +3,8 @@
 (def init-db {:output-channel nil
               :extension-context nil
               :invoked-script nil
-              :disposables []})
+              :disposables []
+              :workspace-root-path nil})
 
 (defonce !app-db (atom init-db))